Created and tested Trace plots
We need to decide where this should go. I currently have it in convergence calculations since it has discrete chains samples passed in. We can easily have this as a plotting function, but we need to make sure discrete chains is saved. Discrete chains is not needed (ie. MH can work perfectly fine), but it is very useful for multi walker MCMC runs.